\docType{data}
\name{ACT_Border}
\alias{ACT_Border}
\title{A polygon of the Australian Capital Territory border}
\format{
An sf data frame with 1 feature and 0 fields.
\describe{
\item{Geometry}{The geometry of the ACT border.}
}
}
\source{
{See Data/ACT_Border.RData in LICENSE.md.}
}
\usage{
data(ACT_Border, package = "STordinal")
}
\description{
A polygon of the Australian Capital Territory (ACT) border with the GDA2020 projection.  
}
\keyword{datasets}
