% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dfBinary}
\alias{dfBinary}
\title{Convert ordinal data frame for binary data analysis}
\usage{
dfBinary(x, yname)
}
\arguments{
\item{x}{dataframe with each row corresponding to a numeric ordinal valued
observation with name \code{yname} with labels 1, 2, ...}

\item{yname}{character, name of ordinal response variable on positive integers}
}
\value{
an expanded dataframe with additional rows for implied zeroes in
binary likelihood. The new response variable will be named \code{y_binary}.
The cut levels are appended with names \code{cut_1} ... \code{cut_k},
where \eqn{k} is the maximum observed category.
Additionally, a column \code{index} is appended that records the
corresponding index of the original ordinal observation vector.
}
\description{
Convert ordinal data frame for binary data analysis
}
