% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSTSOM-package.R
\docType{package}
\name{DSTSOM-package}
\alias{DSTSOM}
\alias{DSTSOM-package}
\title{DSTSOM: Dynamic Spatio-Temporal Sequential Ordinal Models}
\description{
Space time models for ordinal analysis with application to weeds of the 
Australian Capital Territory (ACT) using Dynamic Spatio-Temporal Sequential 
Ordinal Models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csiro-external/DSTSOM}
}

}
\author{
\strong{Maintainer}: Geoffrey R. Hosack \email{geoff.hosack@csiro.au} (\href{https://orcid.org/0000-0002-6462-6817}{ORCID})

Authors:
\itemize{
  \item Wen-Hsi Yang \email{wen-hsi.yang@data61.csiro.au} (\href{https://orcid.org/0000-0001-8236-0082}{ORCID})
  \item Kyana N. Pike \email{kyana.pike@csiro.au} (\href{https://orcid.org/0000-0001-9259-2899}{ORCID})
}

}
\keyword{internal}
