% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_SpatialRandom_consistent}
\alias{is_SpatialRandom_consistent}
\title{Verify if the predicted spatial random effects for the input ordinal data
points at the same location and time point are the same.}
\usage{
is_SpatialRandom_consistent(
  inlaObj = NULL,
  Aproj = NULL,
  DataLocsTime = NULL,
  spdeName = NULL
)
}
\arguments{
\item{inlaObj}{the inla output object of a spatial or spatial-temporal
ordinal model.}

\item{Aproj}{the projection matrix A used to fit the spatial or
spatial-temporal ordinal model which produces the inlaObj.}

\item{DataLocsTime}{the matrix of locations and/or time in the input ordinal
data fitted to a spatial or spatial-temporal model. DataLocsTime should
be a matrix of two columns representing locations of the data fitted to the
spatial model. Otherwise, DataLocsTime should be a matrix of three columns
representing locations and time of the data fitted to the spatial-temporal
model.}

\item{spdeName}{the name of spatial random effect}
}
\value{
a logic value. If TRUE, the predicted spatial random effects are the
same; if FALSE, they are different.
}
\description{
Verify if the predicted spatial random effects for the input ordinal data
points at the same location and time point are the same.
}
