% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_DSTSOM}
\alias{predict_DSTSOM}
\title{Predict marginal probabilities of ordinal levels in space-time.}
\usage{
predict_DSTSOM(
  fm_inla = NULL,
  nSample = 50,
  arrPred = NULL,
  A_s_Pred = NULL,
  INLAseed = 0L
)
}
\arguments{
\item{fm_inla}{spatio-temporal ordinal model fit by INLA. The model fixed
effects include cut levels labelled by cut_1, cut_2, ... cut_k - 1 for k
ordinal levels that correspond to first k - 1 columns in model matrix. The
model formula should include no intercept, e.g., \code{y ~ -1 + cut_1 +
cut_2 + ...}. The name of first argument in inla.spde.make.index is "s",
e.g., \code{inla.spde.make.index('s', n.spde, n.group)}. Please refer to
paper and worked examples in package vignette for further details on INLA
model formulation.}

\item{nSample}{number of Monte Carlo realisations sampled from ordinal model
by INLA}

\item{arrPred}{array of sites (first dimension) by fixed covariates less cut
levels (second dimension) by time levels (e.g., years, third dimension)
conformal to model matrix so that covariate names  match, in order, "fixed"
covariates from INLA model fit excluding cut levels.}

\item{A_s_Pred}{prediction weight matrix output from
\code{INLA::inla.spde.make.A} based on spatial mesh for prediction}

\item{INLAseed}{seed passed to INLA posterior sampling function}
}
\value{
List of 1) MCsummary object that summarises predictive posterior
quantiles by site, ordinal level and year, and 2) a summary matrix maxDev.
The latter is a numerical check to show that max absolute deviation from one
after summation of predicted marginal probabilities by ordinal level across
all realisations of each site and year combination.
}
\description{
Predict marginal probabilities of ordinal levels in space-time.
}
